% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_features.R
\name{get_features}
\alias{get_features}
\title{Generate string distance metrics for all combinations of two vectors}
\usage{
get_features(string1, string2)
}
\arguments{
\item{string1}{A character vector}

\item{string2}{A character vector}
}
\value{
A data frame of rows length(string1) * length(string2) with columns for each string distance metric
}
\description{
Generate string distance metrics for all combinations of two vectors
}
\examples{
get_features(string1 = c("hello", "goodbye"),
             string2 = c("alpha", "beta"))
}
