% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringmatch.R
\name{stringmatch}
\alias{stringmatch}
\title{Generate a string distance score for all combinations of two vectors}
\usage{
stringmatch(string1, string2, trainingset = NULL, feature.importances = FALSE)
}
\arguments{
\item{string1}{A character vector}

\item{string2}{A character vector}

\item{trainingset}{If NULL, use the base model. Otherwise, a data frame with columns A, B, and y corresponding to string1, string2, and a training label}

\item{feature.importances}{If TRUE, returns feature importances.}
}
\value{
A data frame of rows length(string1) * length(string2) with columns for each string distance metric as well as a model-produced score
}
\description{
Generate a string distance score for all combinations of two vectors
}
\examples{
stringmatch(string1 = c("hello", "goodbye"),
             string2 = c("alpha", "beta"))
}
